Move controllers to same namespace as their model
